create_drf_app/templates/basic/{{ cookiecutter.project_name }}/requirements.txt
django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.1